backend/requirements.txt
# Project dependencies
fastapi
uvicorn
google-generativeai
faker
pydantic
python-dotenv
pandas
